+++ src/AvePoint.Elements/Public/Get-AvptProductOverview.ps1
@@ -5,14 +5,15 @@ function Get-AvptProductOverview {
     #>
     [CmdletBinding()]
     param(
-        [Parameter(Mandatory)]
         [string] $CustomerId,
 
-        [Parameter(Mandatory)]
         [int] $ProductType,
 
         [switch] $Raw
     )
+
+    $CustomerId = (Resolve-AvptCustomerSelection -CustomerId $CustomerId).CustomerId
+    $ProductType = Resolve-AvptProductTypeSelection -ProductType $ProductType
 
     $path = "/partner/external/v3/general/customers/$CustomerId/avpt-products/type/$ProductType/overview"
     $item = Invoke-AvptWebRequest -Method Get -Path $path -ScopeBundle 'Common'
@@ -23,4 +24,3 @@ function Get-AvptProductOverview {
 
     ConvertTo-AvptFriendlyObject -InputObject $item -TypeName 'AvePoint.Elements.ProductOverview'
 }
-